requirements.txt
# Core dependencies for web deployment
litellm
deepgram-sdk
python-dotenv
pydantic
streamlit
audio-recorder-streamlit